Restructure code to support multiple networks
