+++ markets/perps-market/subgraph/matchstick.yaml
@@ -1,2 +1,3 @@
 testFolder: tests/
-libsFolder: ../../node_modules/
+libsFolder: ../../../node_modules/
+manifestPath: ./subgraph.optimism-goerli.yaml
