+++ markets/perps-market/subgraph/optimism-goerli/handleAccountCreated.ts
@@ -1,5 +1,5 @@
-import { AccountCreated } from '../generated/PerpsMarketProxy/PerpsMarketProxy';
-import { Account } from '../generated/schema';
+import { AccountCreated } from './generated/PerpsMarketProxy/PerpsMarketProxy';
+import { Account } from './generated/schema';
 
 export function handleAccountCreated(event: AccountCreated): void {
   const id = event.params.accountId.toString();
