+++ markets/perps-market/subgraph/optimism-goerli/handleLockedOiRatioSet.ts
@@ -1,5 +1,5 @@
-import { LockedOiRatioSet } from '../generated/PerpsMarketProxy/PerpsMarketProxy';
-import { Market } from '../generated/schema';
+import { LockedOiRatioSet } from './generated/PerpsMarketProxy/PerpsMarketProxy';
+import { Market } from './generated/schema';
 
 export function handleLockedOiRatioSet(event: LockedOiRatioSet): void {
   const id = event.params.marketId.toString();
