+++ markets/perps-market/subgraph/optimism-goerli/schema.graphql
