+++ markets/perps-market/subgraph/package.json
@@ -4,22 +4,20 @@
   "private": true,
   "version": "1.0.0",
   "scripts": {
-    "subgraphgen": "yarn deployments:optimism-goerli && yarn codegen",
-    "deployments": "yarn deployments:optimism-mainnet && yarn deployments:optimism-goerli",
+    "subgraphgen": "yarn subgraph:deployments && yarn subgraph:codegen && yarn subgraph:build",
+    "subgraph:deployments": "yarn deployments:optimism-goerli",
+    "subgraph:codegen": "yarn codegen:optimism-goerli",
+    "subgraph:build": "yarn build:optimism-goerli",
     "auth": "graph auth --product hosted-service $V3_GRAPH_KEY",
-    "deployments:optimism-mainnet": "echo SKIP || cannon inspect synthetix-perps-market:latest --chain-id 10 --write-deployments ./deployments/optimism-mainnet",
-    "deployments:optimism-goerli": "cannon inspect synthetix-perps-market:latest --chain-id 420 --write-deployments ./deployments/optimism-goerli",
-    "prepare": "cp ./subgraph.base.yaml ./subgraph.yaml",
-    "codegen": "yarn prepare && graph codegen && prettier --write .",
-    "build:optimism-mainnet": "echo SKIP || node ./codegen optimism-mainnet && yarn codegen && graph build --output-dir ./build/optimism-mainnet --network optimism",
-    "deploy:optimism-mainnet": "echo SKIP || graph deploy --output-dir ./build/optimism-mainnet --network optimism --product hosted-service snx-v3/perps-market-optimism-mainnet",
-    "build:optimism-goerli": "node ./codegen optimism-goerli && yarn codegen && graph build --output-dir ./build/optimism-goerli --network optimism-goerli",
-    "deploy:optimism-goerli": "graph deploy --output-dir ./build/optimism-goerli --network optimism-goerli --product hosted-service snx-v3/v3-perps-opt-goerli",
+    "deployments:optimism-goerli": "cannon inspect synthetix-perps-market:latest --chain-id 420 --write-deployments ./optimism-goerli/deployments",
+    "codegen:optimism-goerli": "node ./codegen optimism-goerli && graph codegen subgraph.optimism-goerli.yaml --output-dir=optimism-goerli/generated && prettier --write optimism-goerli/generated",
+    "build:optimism-goerli": "graph build subgraph.optimism-goerli.yaml --output-dir ./build/optimism-goerli --network optimism-goerli",
+    "deploy:optimism-goerli": "graph deploy snx-v3/v3-perps-opt-goerli subgraph.optimism-goerli.yaml --output-dir ./build/optimism-goerli --network optimism-goerli --product hosted-service",
     "test": "echo 'SKIP: No tests' || graph test",
     "coverage": "echo 'SKIP: No tests'",
-    "create-local": "graph create --node http://localhost:8020/ snx-v3/goerli",
-    "remove-local": "graph remove --node http://localhost:8020/ snx-v3/goerli",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 snx-v3/goerli"
+    "create-local": "graph create --node http://localhost:8020/ snx-v3/v3-perps-opt-goerli",
+    "remove-local": "graph remove --node http://localhost:8020/ snx-v3/v3-perps-opt-goerli",
+    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 snx-v3/v3-perps-opt-goerli"
   },
   "devDependencies": {
     "@graphprotocol/graph-cli": "^0.50.1",
