+++ markets/perps-market/subgraph/subgraph.optimism-goerli.yaml
@@ -1,18 +1,22 @@
 specVersion: 0.0.4
 schema:
-  file: ./schema.graphql
+  file: ./optimism-goerli/schema.graphql
 dataSources:
   - kind: ethereum
+    network: optimism-goerli
     name: PerpsMarketProxy
     source:
       abi: PerpsMarketProxy
+      address: "0xf272382cB3BE898A8CdB1A23BE056fA2Fcf4513b"
+      startBlock: 12708889
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
+      file: ./optimism-goerli/index.ts
       abis:
         - name: PerpsMarketProxy
-          file: ./deployments/optimism-goerli/PerpsMarketProxy.json
+          file: ./optimism-goerli/deployments/PerpsMarketProxy.json
       entities:
         - Account
         - Market
@@ -36,26 +40,34 @@ dataSources:
           handler: handleMarketPriceDataUpdated
         - event: FundingParametersSet(indexed uint128,uint256,uint256)
           handler: handleFundingParametersSet
-        - event: LiquidationParametersSet(indexed uint128,uint256,uint256,uint256,uint256,uint256,uint256,uint256)
+        - event: LiquidationParametersSet(indexed
+            uint128,uint256,uint256,uint256,uint256,uint256,uint256,uint256)
           handler: handleLiquidationParametersSet
         - event: LockedOiRatioSet(indexed uint128,uint256)
           handler: handleLockedOiRatioSet
         - event: OrderFeesSet(indexed uint128,uint256,uint256)
           handler: handleOrderFeesSet
-        - event: SettlementStrategyAdded(indexed uint128,(uint8,uint256,uint256,uint256,address,bytes32,string,uint256,uint256,bool),indexed uint256)
+        - event: SettlementStrategyAdded(indexed
+            uint128,(uint8,uint256,uint256,uint256,address,bytes32,string,uint256,uint256,bool),indexed
+            uint256)
           handler: handleSettlementStrategyAdded
         - event: SettlementStrategyEnabled(indexed uint128,uint256,bool)
           handler: handleSettlementStrategyEnabled
-        - event: OrderCommitted(indexed uint128,indexed uint128,uint8,int128,uint256,uint256,uint256,indexed bytes32,address)
+        - event: OrderCommitted(indexed uint128,indexed
+            uint128,uint8,int128,uint256,uint256,uint256,indexed
+            bytes32,address)
           handler: handleOrderCommitted
-        - event: OrderSettled(indexed uint128,indexed uint128,uint256,int256,int256,int128,int128,uint256,uint256,uint256,uint256,indexed bytes32,address)
+        - event: OrderSettled(indexed uint128,indexed
+            uint128,uint256,int256,int256,int128,int128,uint256,uint256,uint256,uint256,indexed
+            bytes32,address)
           handler: handleOrderSettled
-        - event: PreviousOrderExpired(indexed uint128,indexed uint128,int128,uint256,uint256,indexed bytes32)
+        - event: PreviousOrderExpired(indexed uint128,indexed
+            uint128,int128,uint256,uint256,indexed bytes32)
           handler: handlePreviousOrderExpired
         - event: PositionLiquidated(indexed uint128,indexed uint128,uint256,int128)
           handler: handlePositionLiquidated
         - event: AccountLiquidated(indexed uint128,uint256,bool)
           handler: handleAccountLiquidated
-        - event: CollateralModified(indexed uint128,indexed uint128,int256,indexed address)
+        - event: CollateralModified(indexed uint128,indexed uint128,int256,indexed
+            address)
           handler: handleCollateralModified
-      file: ./src/core.ts
